refactor: do not branch out from original loss functions
