+++ src/neural_mesh_simplification/losses/__init__.py
@@ -1,1 +1,1 @@
-from .dgl.combined_loss_dgl import CombinedMeshSimplificationLoss
+from neural_mesh_simplification.losses.combined_loss_dgl import CombinedMeshSimplificationLoss
