+++ src/neural_mesh_simplification/losses/triangle_collision_loss_dgl.py
