+++ tests/losses/test_overlapping_triangles_loss.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from neural_mesh_simplification.losses.dgl.overlapping_triangles_loss_dgl import OverlappingTrianglesLoss
+from neural_mesh_simplification.losses.overlapping_triangles_loss_dgl import OverlappingTrianglesLoss
 
 
 @pytest.fixture
